Refactor: Extract Rule dialog UI into module
Extract the rule window related code in order to simplify and shrink the
diversion rule module.
